Made progress

+++ app/src/main/java/com/example/bulbapp/bulbapplib/adapters/WizLightAdapter.java
@@ -11,34 +11,34 @@ public class WizLightAdapter implements Light {
     }
 
     @Override
-    public void setRGB(byte red, byte green, byte blue) {
+    public void setRGB(int red, int green, int blue) {
         wizLight.getVariables().setR(red);
         wizLight.getVariables().setG(green);
         wizLight.getVariables().setB(blue);
     }
 
     @Override
-    public void setWhite(byte white) {
+    public void setWhite(int white) {
         wizLight.getVariables().setW(white);
     }
 
     @Override
-    public byte getRed() {
+    public int getRed() {
         return wizLight.getVariables().getR();
     }
 
     @Override
-    public byte getGreen() {
+    public int getGreen() {
         return wizLight.getVariables().getG();
     }
 
     @Override
-    public byte getBlue() {
+    public int getBlue() {
         return wizLight.getVariables().getB();
     }
 
     @Override
-    public byte getWhite() {
+    public int getWhite() {
         return wizLight.getVariables().getW();
     }
