+++ app/src/main/java/com/example/bulbapp/bulbapplib/models/Light.java
@@ -1,11 +1,11 @@
 package com.example.bulbapp.bulbapplib.models;
 
 public interface Light {
-    void setRGB(byte red, byte green, byte blue);
-    void setWhite(byte white);
-    byte getRed();
-    byte getGreen();
-    byte getBlue();
-    byte getWhite();
+    void setRGB(int red, int green, int blue);
+    void setWhite(int white);
+    int getRed();
+    int getGreen();
+    int getBlue();
+    int getWhite();
     String getId();
 }
